Added tests for the new code, removed tests for old code and split views tests into small classes and files
