+++ concordia/tests/test_views.py
@@ -1,11 +1,9 @@
 import json
-import sys
 from datetime import date, timedelta
 from unittest.mock import patch
 
 from django import forms
-from django.conf import settings
-from django.contrib.auth.models import AnonymousUser, User
+from django.contrib.auth.models import AnonymousUser
 from django.core.cache import caches
 from django.db.models.signals import post_save
 from django.http import HttpResponse, JsonResponse
@@ -13,7 +11,6 @@ from django.test import (
     Client,
     RequestFactory,
     TestCase,
-    TransactionTestCase,
     override_settings,
 )
 from django.urls import reverse
@@ -22,7 +19,6 @@ from django.utils.timezone import now
 
 from concordia.models import (
     Asset,
-    AssetTranscriptionReservation,
     Campaign,
     ProjectTopic,
     Transcription,
@@ -31,11 +27,8 @@ from concordia.models import (
 from concordia.signals.handlers import on_transcription_save
 from concordia.tasks import (
     campaign_report,
-    delete_old_tombstoned_reservations,
-    expire_inactive_asset_reservations,
-    tombstone_old_active_asset_reservations,
 )
-from concordia.utils import get_anonymous_user, get_or_create_reservation_token
+from concordia.utils import get_anonymous_user
 from concordia.views.accounts import AccountProfileView, registration_rate
 from concordia.views.campaigns import CompletedCampaignListView
 from concordia.views.decorators import reserve_rate, user_cache_control
@@ -43,7 +36,6 @@ from concordia.views.items import FilteredItemDetailView
 from concordia.views.projects import FilteredProjectDetailView
 from concordia.views.rate_limit import ratelimit_view
 from concordia.views.visualizations import VisualizationDataView
-from configuration.models import Configuration
 
 from .utils import (
     CreateTestUsers,
@@ -929,1420 +921,6 @@ class ConcordiaViewTests(CreateTestUsers, JSONAssertMixin, TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, "transcriptions/campaign_report.html")
         self.assertEqual(ctx["projects"].number, 1)
-
-
-@override_settings(
-    RATELIMIT_ENABLE=False, SESSION_ENGINE="django.contrib.sessions.backends.cache"
-)
-class TransactionalViewTests(CreateTestUsers, JSONAssertMixin, TransactionTestCase):
-    def test_asset_reservation(self):
-        """
-        Test the basic Asset reservation process
-        """
-
-        self.login_user()
-        self._asset_reservation_test_payload(self.user.pk)
-
-    def test_asset_reservation_anonymously(self):
-        """
-        Test the basic Asset reservation process as an anonymous user
-        """
-
-        anon_user = get_anonymous_user()
-        self._asset_reservation_test_payload(anon_user.pk, anonymous=True)
-
-    def _asset_reservation_test_payload(self, user_id, anonymous=False):
-        asset = create_asset()
-
-        # Acquire the reservation: 1 acquire
-        # + 1 reservation check
-        # + 1 logging if not anonymous
-        # + 1 session if not anonymous and using a database session engine:
-        expected_update_queries = 2
-        if not anonymous:
-            expected_update_queries += 1  # Added by django-structlog middleware
-            if settings.SESSION_ENGINE.endswith("db"):
-                expected_update_queries += 1  # Added by database session engine
-            # We don't need to add an extra query for accessing request.user
-            # because the django-structlog middleware will do that for non-anonymous
-            expected_acquire_queries = expected_update_queries
-        else:
-            expected_acquire_queries = expected_update_queries + 1
-
-        with self.assertNumQueries(expected_acquire_queries):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-        data = self.assertValidJSON(resp, expected_status=200)
-
-        reservation = AssetTranscriptionReservation.objects.get()
-        self.assertEqual(reservation.reservation_token, data["reservation_token"])
-        self.assertEqual(reservation.asset, asset)
-
-        # Confirm that an update did not change the pk when it updated the timestamp:
-
-        with self.assertNumQueries(expected_update_queries):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-        updated_reservation = AssetTranscriptionReservation.objects.get()
-        self.assertEqual(
-            updated_reservation.reservation_token, data["reservation_token"]
-        )
-        self.assertEqual(updated_reservation.asset, asset)
-        self.assertEqual(reservation.created_on, updated_reservation.created_on)
-        self.assertLess(reservation.created_on, updated_reservation.updated_on)
-
-        # Release the reservation now that we're done:
-        # 1 release
-        # + 1 logging if not anonymous
-        # + 1 session if not anonymous and using a database
-        expected_release_queries = 1
-        if not anonymous:
-            expected_release_queries += 1  # Added by django-structlog middleware
-            if settings.SESSION_ENGINE.endswith("db"):
-                expected_release_queries += 1
-
-        with self.assertNumQueries(expected_release_queries):
-            resp = self.client.post(
-                reverse("reserve-asset", args=(asset.pk,)), data={"release": True}
-            )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertEqual(
-            updated_reservation.reservation_token, data["reservation_token"]
-        )
-
-        self.assertEqual(0, AssetTranscriptionReservation.objects.count())
-
-    def test_asset_reservation_competition(self):
-        """
-        Confirm that two users cannot reserve the same asset at the same time
-        """
-
-        asset = create_asset()
-
-        # We'll reserve the test asset as the anonymous user and then attempt
-        # to edit it after logging in
-
-        # 4 queries =
-        # 1 expiry + 1 acquire + 2 get user ID + 2 get user profile from request
-        with self.assertNumQueries(6):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-        self.assertEqual(200, resp.status_code)
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-
-        # Clear the login session so the reservation_token will be regenerated:
-        self.client.logout()
-        self.login_user()
-
-        # 1 session check + 1 acquire + get user ID from request
-        with self.assertNumQueries(3 if settings.SESSION_ENGINE.endswith("db") else 2):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-        self.assertEqual(409, resp.status_code)
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-
-    def test_asset_reservation_expiration(self):
-        """
-        Simulate an expired reservation which should not cause the request to fail
-        """
-        asset = create_asset()
-
-        stale_reservation = AssetTranscriptionReservation(  # nosec
-            asset=asset, reservation_token="stale"
-        )
-        stale_reservation.full_clean()
-        stale_reservation.save()
-        # Backdate the object as if it happened 31 minutes ago:
-        old_timestamp = now() - timedelta(minutes=31)
-        AssetTranscriptionReservation.objects.update(
-            created_on=old_timestamp, updated_on=old_timestamp
-        )
-
-        expire_inactive_asset_reservations()
-
-        self.login_user()
-
-        # 1 reservation check + 1 acquire + 1 get user ID from request
-        expected_queries = 3
-        if settings.SESSION_ENGINE.endswith("db"):
-            # 1 session check
-            expected_queries += 1
-
-        with self.assertNumQueries(expected_queries):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-        reservation = AssetTranscriptionReservation.objects.get()
-        self.assertEqual(reservation.reservation_token, data["reservation_token"])
-
-    def test_asset_reservation_tombstone(self):
-        """
-        Simulate a tombstoned reservation which should:
-            - return 408 during the tombstone period
-            - during the tombstone period, another user may
-              obtain the reservation but the original user may not
-        """
-        asset = create_asset()
-        self.login_user()
-        request_factory = RequestFactory()
-        request = request_factory.get("/")
-        request.session = {}
-        reservation_token = get_or_create_reservation_token(request)
-
-        session = self.client.session
-        session["reservation_token"] = reservation_token
-        session.save()
-
-        tombstone_reservation = AssetTranscriptionReservation(  # nosec
-            asset=asset, reservation_token=reservation_token
-        )
-        tombstone_reservation.full_clean()
-        tombstone_reservation.save()
-        # Backdate the object as if it was created hours ago,
-        # even if it was recently updated
-        old_timestamp = now() - timedelta(
-            hours=settings.TRANSCRIPTION_RESERVATION_TOMBSTONE_HOURS + 1
-        )
-        current_timestamp = now()
-        AssetTranscriptionReservation.objects.update(
-            created_on=old_timestamp, updated_on=current_timestamp
-        )
-
-        tombstone_old_active_asset_reservations()
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-        reservation = AssetTranscriptionReservation.objects.get()
-        self.assertEqual(reservation.tombstoned, True)
-
-        # 1 session check + 1 reservation check + 1 logging
-        if settings.SESSION_ENGINE.endswith("db"):
-            expected_queries = 3
-        else:
-            expected_queries = 2
-
-        with self.assertNumQueries(expected_queries):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-
-        self.assertEqual(resp.status_code, 408)
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-        reservation = AssetTranscriptionReservation.objects.get()
-        self.assertEqual(reservation.reservation_token, reservation_token)
-
-        self.client.logout()
-
-        # 1 reservation check + 1 acquire + 1 get user ID
-        expected_queries = 3
-        if settings.SESSION_ENGINE.endswith("db"):
-            # + 1 session check
-            expected_queries += 1
-
-        User.objects.create_user(username="anonymous")
-        with self.assertNumQueries(expected_queries):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-
-        self.assertValidJSON(resp, expected_status=200)
-        self.assertEqual(2, AssetTranscriptionReservation.objects.count())
-
-    def test_asset_reservation_tombstone_expiration(self):
-        """
-        Simulate a tombstoned reservation which should expire after
-        the configured period of time, allowing the original user
-        to reserve the asset again
-        """
-        asset = create_asset()
-        self.login_user()
-        request_factory = RequestFactory()
-        request = request_factory.get("/")
-        request.session = {}
-        reservation_token = get_or_create_reservation_token(request)
-
-        session = self.client.session
-        session["reservation_token"] = reservation_token
-        session.save()
-
-        tombstone_reservation = AssetTranscriptionReservation(  # nosec
-            asset=asset, reservation_token=reservation_token
-        )
-        tombstone_reservation.full_clean()
-        tombstone_reservation.save()
-        # Backdate the object as if it was created hours ago
-        # and tombstoned hours ago
-        old_timestamp = now() - timedelta(
-            hours=settings.TRANSCRIPTION_RESERVATION_TOMBSTONE_HOURS
-            + settings.TRANSCRIPTION_RESERVATION_TOMBSTONE_LENGTH_HOURS
-            + 1
-        )
-        not_as_old_timestamp = now() - timedelta(
-            hours=settings.TRANSCRIPTION_RESERVATION_TOMBSTONE_LENGTH_HOURS + 1
-        )
-        AssetTranscriptionReservation.objects.update(
-            created_on=old_timestamp, updated_on=not_as_old_timestamp, tombstoned=True
-        )
-
-        delete_old_tombstoned_reservations()
-        self.assertEqual(0, AssetTranscriptionReservation.objects.count())
-
-        # 1 session check + 1 reservation check + 1 acquire + 1logging
-        if settings.SESSION_ENGINE.endswith("db"):
-            expected_queries = 4
-        else:
-            expected_queries = 3
-
-        with self.assertNumQueries(expected_queries):
-            resp = self.client.post(reverse("reserve-asset", args=(asset.pk,)))
-
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertEqual(1, AssetTranscriptionReservation.objects.count())
-        reservation = AssetTranscriptionReservation.objects.get()
-        self.assertEqual(reservation.reservation_token, data["reservation_token"])
-        self.assertEqual(reservation.tombstoned, False)
-
-    def test_transcription_save(self):
-        asset = create_asset()
-
-        # Test when Turnstile validation failes
-        with patch("concordia.turnstile.fields.TurnstileField.validate") as mock:
-            mock.side_effect = forms.ValidationError(
-                "Testing error", code="invalid_turnstile"
-            )
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)), data={"text": "test"}
-            )
-            data = self.assertValidJSON(resp, expected_status=401)
-            self.assertIn("error", data)
-
-        with patch(
-            "concordia.turnstile.fields.TurnstileField.validate", return_value=True
-        ):
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)), data={"text": "test"}
-            )
-            data = self.assertValidJSON(resp, expected_status=201)
-            self.assertIn("submissionUrl", data)
-
-            # Test attempts to create a second transcription without marking that it
-            # supersedes the previous one:
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)), data={"text": "test"}
-            )
-            data = self.assertValidJSON(resp, expected_status=409)
-            self.assertIn("error", data)
-
-            # If a transcription contains a URL, it should return an error
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={
-                    "text": "http://example.com",
-                    "supersedes": asset.transcription_set.get().pk,
-                },
-            )
-            data = self.assertValidJSON(resp, expected_status=400)
-            self.assertIn("error", data)
-
-            # Unacceptable characters: should return 400 with details
-            with patch(
-                "concordia.views.ajax.find_unacceptable_characters"
-            ) as mock_find:
-                # one violation at line 1, col 2: zero-width space
-                mock_find.return_value = [(1, 2, "\u200b")]
-                resp = self.client.post(
-                    reverse("save-transcription", args=(asset.pk,)),
-                    data={"text": "contains bad char"},
-                )
-            data = self.assertValidJSON(resp, expected_status=400)
-            self.assertIn("error", data)
-            self.assertEqual(data.get("error-code"), "unacceptable_characters")
-            self.assertIn("unacceptable-characters", data)
-            # tuples become lists in JSON
-            self.assertEqual(data["unacceptable-characters"], [[1, 2, "\u200b"]])
-
-            # Test that it correctly works when supersedes is set
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={"text": "test", "supersedes": asset.transcription_set.get().pk},
-            )
-            data = self.assertValidJSON(resp, expected_status=201)
-            self.assertIn("submissionUrl", data)
-
-            # Test that it correctly works when supersedes is set and confirm
-            # ocr_originaed is properly set
-            transcription = asset.transcription_set.order_by("pk").last()
-            transcription.ocr_originated = True
-            transcription.save()
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={
-                    "text": "test",
-                    "supersedes": asset.transcription_set.order_by("pk").last().pk,
-                },
-            )
-            data = self.assertValidJSON(resp, expected_status=201)
-            self.assertIn("submissionUrl", data)
-            new_transcription = asset.transcription_set.order_by("pk").last()
-            self.assertTrue(new_transcription.ocr_originated)
-
-            # We should see an error if you attempt to supersede a transcription
-            # which has already been superseded:
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={
-                    "text": "test",
-                    "supersedes": asset.transcription_set.order_by("pk").first().pk,
-                },
-            )
-            data = self.assertValidJSON(resp, expected_status=409)
-            self.assertIn("error", data)
-
-            # We should get an error if you attempt to supersede a transcription
-            # that doesn't exist
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={
-                    "text": "test",
-                    "supersedes": sys.maxsize,
-                },
-            )
-            data = self.assertValidJSON(resp, expected_status=400)
-            self.assertIn("error", data)
-
-            # We should get an error if you attempt to supersede with
-            # with a pk that is invalid (i.e., a string instead of int)
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={
-                    "text": "test",
-                    "supersedes": "bad-pk",
-                },
-            )
-            data = self.assertValidJSON(resp, expected_status=400)
-            self.assertIn("error", data)
-
-            # A logged in user can take over from an anonymous user:
-            self.login_user()
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)),
-                data={
-                    "text": "test",
-                    "supersedes": asset.transcription_set.order_by("pk").last().pk,
-                },
-            )
-            data = self.assertValidJSON(resp, expected_status=201)
-            self.assertIn("submissionUrl", data)
-
-    def test_anonymous_transcription_submission(self):
-        asset = create_asset()
-        anon = get_anonymous_user()
-
-        transcription = Transcription(asset=asset, user=anon, text="previous entry")
-        transcription.full_clean()
-        transcription.save()
-
-        with patch("concordia.turnstile.fields.TurnstileField.validate") as mock:
-            mock.side_effect = forms.ValidationError(
-                "Testing error", code="invalid_turnstile"
-            )
-            resp = self.client.post(
-                reverse("submit-transcription", args=(transcription.pk,))
-            )
-        data = self.assertValidJSON(resp, expected_status=401)
-        self.assertIn("error", data)
-
-        self.assertFalse(Transcription.objects.filter(submitted__isnull=False).exists())
-
-        with patch(
-            "concordia.turnstile.fields.TurnstileField.validate", return_value=True
-        ):
-            self.client.post(
-                reverse("submit-transcription", args=(transcription.pk,)),
-            )
-            self.assertTrue(
-                Transcription.objects.filter(submitted__isnull=False).exists()
-            )
-
-    def test_transcription_submission(self):
-        asset = create_asset()
-
-        with patch(
-            "concordia.turnstile.fields.TurnstileField.validate", return_value=True
-        ):
-            resp = self.client.post(
-                reverse("save-transcription", args=(asset.pk,)), data={"text": "test"}
-            )
-        data = self.assertValidJSON(resp, expected_status=201)
-
-        transcription = Transcription.objects.get()
-        self.assertIsNone(transcription.submitted)
-
-        with patch(
-            "concordia.turnstile.fields.TurnstileField.validate", return_value=True
-        ):
-            resp = self.client.post(
-                reverse("submit-transcription", args=(transcription.pk,))
-            )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertIn("id", data)
-        self.assertEqual(data["id"], transcription.pk)
-
-        transcription = Transcription.objects.get()
-        self.assertTrue(transcription.submitted)
-
-    def test_stale_transcription_submission(self):
-        asset = create_asset()
-
-        anon = get_anonymous_user()
-
-        t1 = Transcription(asset=asset, user=anon, text="test")
-        t1.full_clean()
-        t1.save()
-
-        t2 = Transcription(asset=asset, user=anon, text="test", supersedes=t1)
-        t2.full_clean()
-        t2.save()
-
-        with patch(
-            "concordia.turnstile.fields.TurnstileField.validate", return_value=True
-        ):
-            resp = self.client.post(reverse("submit-transcription", args=(t1.pk,)))
-            data = self.assertValidJSON(resp, expected_status=400)
-            self.assertIn("error", data)
-
-    def test_transcription_review(self):
-        asset = create_asset()
-
-        anon = get_anonymous_user()
-
-        t1 = Transcription(asset=asset, user=anon, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        self.login_user()
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "foobar"}
-        )
-        data = self.assertValidJSON(resp, expected_status=400)
-        self.assertIn("error", data)
-
-        self.assertEqual(
-            1, Transcription.objects.filter(pk=t1.pk, accepted__isnull=True).count()
-        )
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "accept"}
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-
-        self.assertEqual(
-            1, Transcription.objects.filter(pk=t1.pk, accepted__isnull=False).count()
-        )
-
-    def test_transcription_review_rate_limit(self):
-        for cache in caches.all():
-            cache.clear()
-        anon = get_anonymous_user()
-        self.login_user()
-        try:
-            config = Configuration.objects.get(key="review_rate_limit")
-            config.value = "4"
-            config.data_type = Configuration.DataType.NUMBER
-            config.save()
-        except Configuration.DoesNotExist:
-            Configuration.objects.create(
-                key="review_rate_limit",
-                value="4",
-                data_type=Configuration.DataType.NUMBER,
-            )
-
-        Configuration.objects.get_or_create(
-            key="review_rate_limit_popup_message",
-            defaults={
-                "value": "Test message",
-                "data_type": Configuration.DataType.HTML,
-            },
-        )
-        Configuration.objects.get_or_create(
-            key="review_rate_limit_popup_title",
-            defaults={
-                "value": "Test message",
-                "data_type": Configuration.DataType.HTML,
-            },
-        )
-        Configuration.objects.get_or_create(
-            key="review_rate_limit_banner_message",
-            defaults={
-                "value": "Test message",
-                "data_type": Configuration.DataType.HTML,
-            },
-        )
-
-        asset = create_asset()
-        t1 = create_transcription(user=anon, asset=asset)
-        t2 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-2")
-        )
-        t3 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-3")
-        )
-        t4 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-4")
-        )
-        t5 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-5")
-        )
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t2.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t3.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t4.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t5.pk,)), data={"action": "accept"}
-        )
-        data = self.assertValidJSON(resp, expected_status=429)
-        self.assertIn("error", data)
-
-    def test_transcription_review_rate_limit_superuser(self):
-        for cache in caches.all():
-            cache.clear()
-        anon = get_anonymous_user()
-        self.user = self.create_super_user()
-        self.login_user()
-        try:
-            config = Configuration.objects.get(key="review_rate_limit")
-            config.value = "4"
-            config.data_type = Configuration.DataType.NUMBER
-            config.save()
-        except Configuration.DoesNotExist:
-            Configuration.objects.create(
-                key="review_rate_limit",
-                value="4",
-                data_type=Configuration.DataType.NUMBER,
-            )
-
-        Configuration.objects.get_or_create(
-            key="review_rate_limit_popup_message",
-            defaults={
-                "value": "Test message",
-                "data_type": Configuration.DataType.HTML,
-            },
-        )
-        Configuration.objects.get_or_create(
-            key="review_rate_limit_popup_title",
-            defaults={
-                "value": "Test message",
-                "data_type": Configuration.DataType.HTML,
-            },
-        )
-        Configuration.objects.get_or_create(
-            key="review_rate_limit_banner_message",
-            defaults={
-                "value": "Test message",
-                "data_type": Configuration.DataType.HTML,
-            },
-        )
-
-        asset = create_asset()
-        t1 = create_transcription(user=anon, asset=asset)
-        t2 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-2")
-        )
-        t3 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-3")
-        )
-        t4 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-4")
-        )
-        t5 = create_transcription(
-            user=anon, asset=create_asset(item=asset.item, slug="test-asset-5")
-        )
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t2.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t3.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t4.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t5.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-    def test_transcription_review_asset_status_updates(self):
-        """
-        Confirm that the Asset.transcription_status field is correctly updated
-        throughout the review process
-        """
-        asset = create_asset()
-
-        anon = get_anonymous_user()
-
-        # We should see NOT_STARTED only when no transcription records exist:
-        self.assertEqual(asset.transcription_set.count(), 0)
-        self.assertEqual(
-            Asset.objects.get(pk=asset.pk).transcription_status,
-            TranscriptionStatus.NOT_STARTED,
-        )
-
-        t1 = Transcription(asset=asset, user=anon, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        self.assertEqual(
-            Asset.objects.get(pk=asset.pk).transcription_status,
-            TranscriptionStatus.SUBMITTED,
-        )
-
-        # “Login” so we can review the anonymous transcription:
-        self.login_user()
-
-        self.assertEqual(
-            1, Transcription.objects.filter(pk=t1.pk, accepted__isnull=True).count()
-        )
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "reject"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        # After rejecting a transcription, the asset status should be reset to
-        # in-progress:
-        self.assertEqual(
-            1,
-            Transcription.objects.filter(
-                pk=t1.pk, accepted__isnull=True, rejected__isnull=False
-            ).count(),
-        )
-        self.assertEqual(
-            Asset.objects.get(pk=asset.pk).transcription_status,
-            TranscriptionStatus.IN_PROGRESS,
-        )
-
-        # We'll simulate a second attempt:
-
-        t2 = Transcription(
-            asset=asset, user=anon, text="test", submitted=now(), supersedes=t1
-        )
-        t2.full_clean()
-        t2.save()
-
-        self.assertEqual(
-            Asset.objects.get(pk=asset.pk).transcription_status,
-            TranscriptionStatus.SUBMITTED,
-        )
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t2.pk,)), data={"action": "accept"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-
-        self.assertEqual(
-            1, Transcription.objects.filter(pk=t2.pk, accepted__isnull=False).count()
-        )
-        self.assertEqual(
-            Asset.objects.get(pk=asset.pk).transcription_status,
-            TranscriptionStatus.COMPLETED,
-        )
-
-    def test_transcription_disallow_self_review(self):
-        asset = create_asset()
-
-        self.login_user()
-
-        t1 = Transcription(asset=asset, user=self.user, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "accept"}
-        )
-        data = self.assertValidJSON(resp, expected_status=400)
-        self.assertIn("error", data)
-        self.assertEqual("You cannot accept your own transcription", data["error"])
-
-    def test_transcription_allow_self_reject(self):
-        asset = create_asset()
-
-        self.login_user()
-
-        t1 = Transcription(asset=asset, user=self.user, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "reject"}
-        )
-        self.assertValidJSON(resp, expected_status=200)
-        self.assertEqual(
-            Asset.objects.get(pk=asset.pk).transcription_status,
-            TranscriptionStatus.IN_PROGRESS,
-        )
-        self.assertEqual(Transcription.objects.get(pk=t1.pk).reviewed_by, self.user)
-
-    def test_transcription_double_review(self):
-        asset = create_asset()
-
-        anon = get_anonymous_user()
-
-        t1 = Transcription(asset=asset, user=anon, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        self.login_user()
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "accept"}
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-
-        resp = self.client.post(
-            reverse("review-transcription", args=(t1.pk,)), data={"action": "reject"}
-        )
-        data = self.assertValidJSON(resp, expected_status=400)
-        self.assertIn("error", data)
-        self.assertEqual("This transcription has already been reviewed", data["error"])
-
-    def test_anonymous_tag_submission(self):
-        """Confirm that anonymous users cannot submit tags"""
-        asset = create_asset()
-        submit_url = reverse("submit-tags", kwargs={"asset_pk": asset.pk})
-
-        resp = self.client.post(submit_url, data={"tags": ["foo", "bar"]})
-        self.assertRedirects(resp, "%s?next=%s" % (reverse("login"), submit_url))
-
-    def test_tag_submission(self):
-        asset = create_asset()
-
-        self.login_user()
-
-        test_tags = ["foo", "bar"]
-
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": test_tags},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertIn("user_tags", data)
-        self.assertIn("all_tags", data)
-
-        self.assertEqual(sorted(test_tags), data["user_tags"])
-        self.assertEqual(sorted(test_tags), data["all_tags"])
-
-    def test_invalid_tag_submission(self):
-        asset = create_asset()
-
-        self.login_user()
-
-        test_tags = ["foo", "bar"]
-
-        with patch("concordia.models.Tag.full_clean") as mock:
-            mock.side_effect = forms.ValidationError("Testing error")
-            resp = self.client.post(
-                reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-                data={"tags": test_tags},
-            )
-            data = self.assertValidJSON(resp, expected_status=400)
-            self.assertIn("error", data)
-
-    def test_tag_submission_with_diacritics(self):
-        asset = create_asset()
-
-        self.login_user()
-
-        test_tags = ["Café", "château", "señor", "façade"]
-
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": test_tags},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertIn("user_tags", data)
-        self.assertIn("all_tags", data)
-
-        self.assertEqual(sorted(test_tags), data["user_tags"])
-        self.assertEqual(sorted(test_tags), data["all_tags"])
-
-    def test_tag_submission_with_multiple_users(self):
-        asset = create_asset()
-        self.login_user()
-
-        test_tags = ["foo", "bar"]
-
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": test_tags},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertIn("user_tags", data)
-        self.assertIn("all_tags", data)
-
-        self.assertEqual(sorted(test_tags), data["user_tags"])
-        self.assertEqual(sorted(test_tags), data["all_tags"])
-
-    def test_duplicate_tag_submission(self):
-        """Confirm that tag values cannot be duplicated"""
-        asset = create_asset()
-
-        self.login_user()
-
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": ["foo", "bar", "baaz"]},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-
-        second_user = self.create_test_user(
-            username="second_tester", email="second_tester@example.com"
-        )
-        self.client.login(username=second_user.username, password=second_user._password)
-
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": ["foo", "bar", "quux"]},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-
-        # Even though the user submitted (through some horrible bug) duplicate
-        # values, they should not be stored:
-        self.assertEqual(["bar", "foo", "quux"], data["user_tags"])
-        # Users are allowed to delete other users' tags, so since the second
-        # user didn't send the "baaz" tag, it was removed
-        self.assertEqual(["bar", "foo", "quux"], data["all_tags"])
-
-    def test_tag_deletion(self):
-        asset = create_asset()
-        self.login_user()
-
-        initial_tags = ["foo", "bar"]
-        self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": initial_tags},
-        )
-        updated_tags = [
-            "foo",
-        ]
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": updated_tags},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertIn("user_tags", data)
-        self.assertIn("all_tags", data)
-
-        self.assertCountEqual(updated_tags, data["user_tags"])
-        self.assertCountEqual(updated_tags, data["all_tags"])
-
-    def test_tag_deletion_with_multiple_users(self):
-        asset = create_asset()
-        self.login_user("first_user")
-        initial_tags = ["foo", "bar"]
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": initial_tags},
-        )
-        self.assertIn(
-            "first_user",
-            asset.userassettagcollection_set.values().values_list(
-                "user__username", flat=True
-            ),
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-        self.assertIn("user_tags", data)
-        self.assertIn("all_tags", data)
-        self.assertCountEqual(initial_tags, data["user_tags"])
-        self.assertCountEqual(initial_tags, data["all_tags"])
-
-        self.client.logout()
-
-        second_user = self.create_test_user("second_user")
-        self.client.login(username=second_user.username, password=second_user._password)
-        updated_tags = [
-            "foo",
-        ]
-        resp = self.client.post(
-            reverse("submit-tags", kwargs={"asset_pk": asset.pk}),
-            data={"tags": updated_tags},
-        )
-        data = self.assertValidJSON(resp, expected_status=200)
-
-        self.assertIn(
-            "second_user",
-            asset.userassettagcollection_set.values().values_list(
-                "user__username", flat=True
-            ),
-        )
-        self.assertEqual(asset.userassettagcollection_set.count(), 2)
-        self.assertEqual(
-            User.objects.filter(userassettagcollection__asset=asset).count(), 2
-        )
-        self.assertIn("user_tags", data)
-        self.assertIn("all_tags", data)
-        self.assertCountEqual(updated_tags, data["user_tags"])
-
-    def test_find_next_transcribable_no_campaign(self):
-        # Test case where there are no transcribable assets
-        resp = self.client.get(reverse("redirect-to-next-transcribable-asset"))
-        self.assertRedirects(resp, expected_url="/")
-
-        asset1 = create_asset(slug="test-asset-1")
-        asset2 = create_asset(item=asset1.item, slug="test-asset-2")
-        campaign = asset1.item.project.campaign
-
-        resp = self.client.get(reverse("redirect-to-next-transcribable-asset"))
-        self.assertRedirects(resp, expected_url=asset1.get_absolute_url())
-
-        # Configure next transcription campaign for tests below
-        campaign.next_transcription_campaign = True
-        campaign.save()
-
-        # Test when next transcribable campaign doesn't exist and there
-        # are no other campaigns/assets
-        with patch("concordia.models.Campaign.objects.get") as mock:
-            mock.side_effect = IndexError
-            response = self.client.get(reverse("redirect-to-next-transcribable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-        # Test case when a campaign is configured to be default next transcribable
-        response = self.client.get(reverse("redirect-to-next-transcribable-asset"))
-        self.assertRedirects(response, expected_url=asset2.get_absolute_url())
-
-        # Test when next transcribable campaign has not transcribable assets
-        asset1.delete()
-        asset2.delete()
-        response = self.client.get(reverse("redirect-to-next-transcribable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-        # Test when next transcription campaign has no transcribable assets
-        # and other campaigns exist and have no transcribable assets
-        create_campaign(slug="test-campaign-2")
-        response = self.client.get(reverse("redirect-to-next-transcribable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-    def test_find_next_reviewable_no_campaign(self):
-        user = self.create_user("test-user")
-        anon = get_anonymous_user()
-
-        # Test case where there are no reviewable assets
-        response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-        asset1 = create_asset(slug="test-asset-1", title="Test Asset 1")
-        asset2 = create_asset(
-            item=asset1.item, slug="test-asset-2", title="Test Asset 2"
-        )
-        asset3 = create_asset(
-            item=asset1.item, slug="test-asset-3", title="Test Asset 3"
-        )
-        campaign = asset1.item.project.campaign
-
-        t1 = Transcription(asset=asset1, user=user, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        t2 = Transcription(asset=asset2, user=anon, text="test", submitted=now())
-        t2.full_clean()
-        t2.save()
-
-        t3 = Transcription(asset=asset3, user=anon, text="test", submitted=now())
-        t3.full_clean()
-        t3.save()
-
-        response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url=asset1.get_absolute_url())
-
-        # Test logged in user (this creates a new user)
-        # asset1 is no longer available due to the request above reserving it
-        self.login_user()
-        response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url=asset2.get_absolute_url())
-
-        # Configure campaign to be next review cmpaign for tests below
-        campaign.next_review_campaign = True
-        campaign.save()
-
-        # Test when next reviewable campaign doesn't exist and there
-        # are no other campaigns/assets
-        with patch("concordia.models.Campaign.objects.get") as mock:
-            mock.side_effect = IndexError
-            response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-        # Test case when a campaign is configured to be default next reviewable
-        response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url=asset3.get_absolute_url())
-
-        # Test when next reviewable campaign has no reviewable assets
-        asset1.delete()
-        asset2.delete()
-        response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-        # Test when next reviewable campaign has no reviewable assets
-        # and other campaigns exist and have no reviewable assets
-        create_campaign(slug="test-campaign-2")
-        response = self.client.get(reverse("redirect-to-next-reviewable-asset"))
-        self.assertRedirects(response, expected_url="/")
-
-    def test_find_next_transcribable_campaign(self):
-        asset1 = create_asset(slug="test-asset-1")
-        asset2 = create_asset(item=asset1.item, slug="test-asset-2")
-        campaign = asset1.item.project.campaign
-
-        # Anonymous user test
-        resp = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                kwargs={"campaign_slug": campaign.slug},
-            )
-        )
-        self.assertRedirects(resp, expected_url=asset1.get_absolute_url())
-
-        # Authenticated user test
-        self.login_user()
-        resp = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                kwargs={"campaign_slug": campaign.slug},
-            )
-        )
-        self.assertRedirects(resp, expected_url=asset2.get_absolute_url())
-
-    def test_find_next_transcribable_topic(self):
-        asset1 = create_asset(slug="test-asset-1")
-        asset2 = create_asset(item=asset1.item, slug="test-asset-2")
-        project = asset1.item.project
-        topic = create_topic(project=project)
-
-        # Anonymous user test
-        resp = self.client.get(
-            reverse(
-                "redirect-to-next-transcribable-topic-asset",
-                kwargs={"topic_slug": topic.slug},
-            )
-        )
-        self.assertRedirects(resp, expected_url=asset1.get_absolute_url())
-
-        # Authenticated user test
-        self.login_user()
-        resp = self.client.get(
-            reverse(
-                "redirect-to-next-transcribable-topic-asset",
-                kwargs={"topic_slug": topic.slug},
-            )
-        )
-        self.assertRedirects(resp, expected_url=asset2.get_absolute_url())
-
-    def test_find_next_reviewable_campaign(self):
-        anon = get_anonymous_user()
-
-        asset1 = create_asset(slug="test-review-asset-1", title="Test Asset 1")
-        asset2 = create_asset(
-            item=asset1.item, slug="test-review-asset-2", title="Test Asset 2"
-        )
-
-        t1 = Transcription(asset=asset1, user=anon, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        t2 = Transcription(asset=asset2, user=anon, text="test", submitted=now())
-        t2.full_clean()
-        t2.save()
-
-        campaign = asset1.item.project.campaign
-
-        # Anonymous user test
-        response = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-reviewable-campaign-asset",
-                kwargs={"campaign_slug": campaign.slug},
-            )
-        )
-        self.assertRedirects(response, expected_url=asset1.get_absolute_url())
-
-        # Authenticated user test
-        # asset1 is no longer available since the previous request reserved it
-        self.login_user()
-        response = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-reviewable-campaign-asset",
-                kwargs={"campaign_slug": campaign.slug},
-            )
-        )
-        self.assertRedirects(response, expected_url=asset2.get_absolute_url())
-
-    def test_find_next_reviewable_topic(self):
-        anon = get_anonymous_user()
-
-        asset1 = create_asset(slug="test-review-asset-1")
-        asset2 = create_asset(item=asset1.item, slug="test-review-asset-2")
-        project = asset1.item.project
-        topic = create_topic(project=project)
-
-        t1 = Transcription(asset=asset1, user=anon, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        t2 = Transcription(asset=asset2, user=anon, text="test", submitted=now())
-        t2.full_clean()
-        t2.save()
-
-        # Anonymous user test
-        response = self.client.get(
-            reverse(
-                "redirect-to-next-reviewable-topic-asset",
-                kwargs={"topic_slug": topic.slug},
-            )
-        )
-        self.assertRedirects(response, expected_url=asset1.get_absolute_url())
-
-        # Authenticated user test
-        # We expect that asset1 is no longer available. Even though
-        # anonymous users can't reserve assets for review, we still will
-        # have removed the asset from the NextReviewableTopicAsset table
-        # to ensure two users don't receive the same asset
-        self.login_user()
-        response = self.client.get(
-            reverse(
-                "redirect-to-next-reviewable-topic-asset",
-                kwargs={"topic_slug": topic.slug},
-            )
-        )
-        self.assertRedirects(response, expected_url=asset2.get_absolute_url())
-
-    def test_find_next_reviewable_unlisted_campaign(self):
-        anon = get_anonymous_user()
-
-        unlisted_campaign = create_campaign(
-            slug="campaign-transcribe-redirect-unlisted",
-            title="Test Unlisted Review Redirect Campaign",
-            unlisted=True,
-        )
-        unlisted_project = create_project(
-            title="Unlisted Project",
-            slug="unlisted-project",
-            campaign=unlisted_campaign,
-        )
-        unlisted_item = create_item(
-            title="Unlisted Item",
-            item_id="unlisted-item",
-            item_url="https://blah.com/unlisted-item",
-            project=unlisted_project,
-        )
-
-        asset1 = create_asset(slug="test-asset-1", item=unlisted_item)
-        asset2 = create_asset(item=asset1.item, slug="test-asset-2")
-
-        t1 = Transcription(asset=asset1, user=anon, text="test", submitted=now())
-        t1.full_clean()
-        t1.save()
-
-        t2 = Transcription(asset=asset2, user=anon, text="test", submitted=now())
-        t2.full_clean()
-        t2.save()
-
-        response = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-reviewable-campaign-asset",
-                kwargs={"campaign_slug": unlisted_campaign.slug},
-            )
-        )
-
-        self.assertRedirects(response, expected_url=asset1.get_absolute_url())
-
-    def test_find_next_transcribable_unlisted_campaign(self):
-        unlisted_campaign = create_campaign(
-            slug="campaign-transcribe-redirect-unlisted",
-            title="Test Unlisted Transcribe Redirect Campaign",
-            unlisted=True,
-        )
-        unlisted_project = create_project(
-            title="Unlisted Project",
-            slug="unlisted-project",
-            campaign=unlisted_campaign,
-        )
-        unlisted_item = create_item(
-            title="Unlisted Item",
-            item_id="unlisted-item",
-            item_url="https://blah.com/unlisted-item",
-            project=unlisted_project,
-        )
-
-        asset1 = create_asset(slug="test-asset-1", item=unlisted_item)
-        create_asset(item=asset1.item, slug="test-asset-2")
-
-        response = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                kwargs={"campaign_slug": unlisted_campaign.slug},
-            )
-        )
-
-        self.assertRedirects(response, expected_url=asset1.get_absolute_url())
-
-    def test_find_next_transcribable_single_asset(self):
-        asset = create_asset()
-        campaign = asset.item.project.campaign
-
-        resp = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                kwargs={"campaign_slug": campaign.slug},
-            )
-        )
-
-        self.assertRedirects(resp, expected_url=asset.get_absolute_url())
-
-    def test_find_next_transcribable_in_singleton_campaign(self):
-        asset = create_asset(transcription_status=TranscriptionStatus.SUBMITTED)
-        campaign = asset.item.project.campaign
-
-        resp = self.client.get(
-            reverse(
-                "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                kwargs={"campaign_slug": campaign.slug},
-            )
-        )
-
-        self.assertRedirects(resp, expected_url=reverse("homepage"))
-
-    def test_find_next_transcribable_project_redirect(self):
-        asset = create_asset(transcription_status=TranscriptionStatus.SUBMITTED)
-        project = asset.item.project
-        campaign = project.campaign
-
-        resp = self.client.get(
-            "%s?project=%s"
-            % (
-                reverse(
-                    "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                    kwargs={"campaign_slug": campaign.slug},
-                ),
-                project.slug,
-            )
-        )
-
-        self.assertRedirects(resp, expected_url=reverse("homepage"))
-
-    def test_find_next_transcribable_hierarchy(self):
-        """Confirm that find-next-page selects assets in the expected order"""
-
-        asset = create_asset()
-        item = asset.item
-        project = item.project
-        campaign = project.campaign
-
-        asset_in_item = create_asset(item=item, slug="test-asset-in-same-item")
-        in_progress_asset_in_item = create_asset(
-            item=item,
-            slug="inprogress-asset-in-same-item",
-            transcription_status=TranscriptionStatus.IN_PROGRESS,
-        )
-
-        asset_in_project = create_asset(
-            item=create_item(project=project, item_id="other-item-in-same-project"),
-            title="test-asset-in-same-project",
-        )
-
-        asset_in_campaign = create_asset(
-            item=create_item(
-                project=create_project(campaign=campaign, title="other project"),
-                title="item in other project",
-            ),
-            slug="test-asset-in-same-campaign",
-        )
-
-        # Now that we have test assets we'll see what find-next-page gives us as
-        # successive test records are marked as submitted and thus ineligible.
-        # The expected ordering is that it will favor moving forward (i.e. not
-        # landing you on the same asset unless that's the only one available),
-        # and will keep you closer to the asset you started from (i.e. within
-        # the same item or project in that order).
-
-        self.assertRedirects(
-            self.client.get(
-                reverse(
-                    "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                    kwargs={"campaign_slug": campaign.slug},
-                ),
-                {"project": project.slug, "item": item.item_id, "asset": asset.pk},
-            ),
-            asset_in_item.get_absolute_url(),
-        )
-
-        asset_in_item.transcription_status = TranscriptionStatus.SUBMITTED
-        asset_in_item.save()
-        AssetTranscriptionReservation.objects.all().delete()
-
-        self.assertRedirects(
-            self.client.get(
-                reverse(
-                    "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                    kwargs={"campaign_slug": campaign.slug},
-                ),
-                {"project": project.slug, "item": item.item_id, "asset": asset.pk},
-            ),
-            asset_in_project.get_absolute_url(),
-        )
-
-        asset_in_project.transcription_status = TranscriptionStatus.SUBMITTED
-        asset_in_project.save()
-        AssetTranscriptionReservation.objects.all().delete()
-
-        self.assertRedirects(
-            self.client.get(
-                reverse(
-                    "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                    kwargs={"campaign_slug": campaign.slug},
-                ),
-                {"project": project.slug, "item": item.item_id, "asset": asset.pk},
-            ),
-            asset_in_campaign.get_absolute_url(),
-        )
-
-        asset_in_campaign.transcription_status = TranscriptionStatus.SUBMITTED
-        asset_in_campaign.save()
-        AssetTranscriptionReservation.objects.all().delete()
-
-        self.assertRedirects(
-            self.client.get(
-                reverse(
-                    "transcriptions:redirect-to-next-transcribable-campaign-asset",
-                    kwargs={"campaign_slug": campaign.slug},
-                ),
-                {"project": project.slug, "item": item.item_id, "asset": asset.pk},
-            ),
-            in_progress_asset_in_item.get_absolute_url(),
-        )
-
-    def tearDown(self):
-        # We'll test the signal handler separately
-        post_save.connect(on_transcription_save, sender=Transcription)
 
 
 class UserCacheControlTest(CreateTestUsers, TestCase):
